Add files via upload
this file should not be merged to avoid the previous mistake. this is the gif I used for the let's go page, instead of using the res folder. if this can't be used to update the already existing  images folder in assets, it should be ignored and code changed to avoid error.